backend/requirements.txt
alembic==1.14.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.8.0
asttokens==2.4.1
attrs==24.2.0
bcrypt==4.2.1
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
charts==0.4.6
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==43.0.1
debugpy==1.8.6
decorator==5.1.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.8
fastapi-cli==0.0.7
firebase-admin==6.5.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.1.0
narwhals==1.9.2
nest-asyncio==1.6.0
numpy==2.1.2
orjson==3.10.15
packaging==24.1
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
patsy==0.5.6
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
plotly-express==0.4.1
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.8.0
pydantic_core==2.27.2
pydeck==0.9.1
pygame==2.6.1
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2024.2
pywin32==307
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rich==13.9.2
rich-toolkit==0.13.2
rpds-py==0.20.0
rsa==4.9
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.45.3
statsmodels==0.14.4
streamlit==1.39.0
tenacity==9.0.0
timedelta==2020.12.3
toml==0.10.2
tornado==6.4.1
traitlets==5.14.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
watchdog==5.0.3
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0
